refactor: restrict domain binding injection
Constraint: Project bindings must be minted by trusted host factories and remain descriptor-bound for each wrapper turn.
Rejected: Retain direct HostProjectBinding injection for tests | It bypasses the production host lifecycle and obscures descriptor ownership.
Confidence: high
Scope-risk: moderate
Directive: Keep wrapper domain-context entrypoints factory-only and split growing project-context tests before they exceed the module limit.
Tested: 35 focused domain binding, privacy, lifecycle, diagnostics, and session tests; Ruff; compileall; static discovery count; signature audit; git diff check.
Not-tested: Full suite awaits migration of direct injection calls owned by the wrapper contract and golden-example lane.
