+++ tests/test_wrapper_sessions.py
@@ -51,7 +51,7 @@ from omh.wrapper_sessions import (
 class DomainContextSessionBindingTests(unittest.TestCase):
     def test_each_turn_uses_a_fresh_host_binding_without_cross_project_reuse(self) -> None:
         from omh.workflows.domain_project_context import bind_plugin_project
-        from test_domain_project_context import _domain_store
+        from domain_project_context_helpers import domain_store as _domain_store
         from test_domain_routing_context import _approve_profile, _repository
 
         with TemporaryDirectory() as tmp:
@@ -63,22 +63,24 @@ class DomainContextSessionBindingTests(unittest.TestCase):
             first_paths = resolve_paths(first_root / ".omh", first_root / ".hermes")
             second_paths = resolve_paths(second_root / ".omh", second_root / ".hermes")
 
-            with bind_plugin_project({"project_root": str(first_root)}) as first_binding:
-                first = create_or_resume_wrapper_session(
-                    first_paths,
-                    "something feels off",
-                    source="discord",
-                    source_metadata={"source_event_id": "m1", "channel_ref": "c1"},
-                    _host_project_binding=first_binding,
-                )
-            with bind_plugin_project({"project_root": str(second_root)}) as second_binding:
-                second = create_or_resume_wrapper_session(
-                    second_paths,
-                    "something feels off",
-                    source="discord",
-                    source_metadata={"source_event_id": "m1", "channel_ref": "c1"},
-                    _host_project_binding=second_binding,
-                )
+            first = create_or_resume_wrapper_session(
+                first_paths,
+                "something feels off",
+                source="discord",
+                source_metadata={"source_event_id": "m1", "channel_ref": "c1"},
+                _host_project_binding_factory=lambda: bind_plugin_project(
+                    {"project_root": str(first_root)}
+                ),
+            )
+            second = create_or_resume_wrapper_session(
+                second_paths,
+                "something feels off",
+                source="discord",
+                source_metadata={"source_event_id": "m1", "channel_ref": "c1"},
+                _host_project_binding_factory=lambda: bind_plugin_project(
+                    {"project_root": str(second_root)}
+                ),
+            )
 
             self.assertIn("domain_routing_context", first["interaction"])
             self.assertNotIn("domain_routing_context", second["interaction"])
